+++ src/app/config/tva/tva.component.ts
@@ -1,14 +1,14 @@
 import { Component, OnInit } from '@angular/core';
-import { TvaService } from '../../services/tva.service'
-import { TvaI} from '../../interfaces/TvaI'
-import { AlertService } from '../../rootComponent/comfirm-dialog/alert.service';
-import { MessageService } from '../../rootComponent/messages/message.service'
-import { MessageI } from '../../interfaces/MessageI'
-import { ExpireService } from '../../services/expire.service';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { take } from 'rxjs/operators';
-import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { MessageI } from '../../interfaces/MessageI';
+import { TvaI } from '../../interfaces/TvaI';
+import { AlertService } from '../../rootComponent/comfirm-dialog/alert.service';
+import { MessageService } from '../../rootComponent/messages/message.service';
+import { ExpireService } from '../../services/expire.service';
+import { TvaService } from '../../services/tva.service';
 
-declare var bootstrap:any;
+declare var bootstrap: any;
 
 @Component({
   selector: 'app-tva',
@@ -18,28 +18,26 @@ declare var bootstrap:any;
 export class TvaComponent implements OnInit {
   tvas: TvaI[];
   selectedTva: TvaI;
-  
-  addForm= new FormGroup({
-    taux:new FormControl('',Validators.required),
 
+  addForm = new FormGroup({
+    taux: new FormControl('', Validators.required),
   });
 
   updateForm = new FormGroup({
-    taux:new FormControl('',Validators.required),
+    taux: new FormControl('', Validators.required),
   });
 
-  updateModal:any;
-  addModal:any;
+  updateModal: any;
+  addModal: any;
 
 
-  onSelect(tva:TvaI)
-  {
-    this.selectedTva=tva;
+  onSelect(tva: TvaI) {
+    this.selectedTva = tva;
   }
- 
-  addData(): void{
-    var taux=this.addForm.get("taux").value;
-    const message:MessageI={content:'La TVA à été rajouté',level:'Info'}
+
+  addData(): void {
+    var taux = this.addForm.get("taux").value;
+    const message: MessageI = { content: 'La TVA à été rajouté', level: 'Info' }
     if (!taux) { return; }
     this.tvaService.addTva({ taux } as TvaI)
       .pipe(take(1)).subscribe(tva => {
@@ -48,48 +46,47 @@ export class TvaComponent implements OnInit {
         this.messageService.add(message);
         this.addForm.reset();
         this.addModal.hide();
-      });    
+      });
   }
 
-  delete(tva:TvaI):void {
-    const message:MessageI={content:'L\'élément à été supprimé',level:'Attention'}
+  delete(tva: TvaI): void {
+    const message: MessageI = { content: 'L\'élément à été supprimé', level: 'Attention' }
     let that = this;
-    this.alertService.confirmThis("Êtes-vous sur de vouloir supprimer la tva ?",function(){
-    that.tvaService.deleteTva(tva).pipe(take(1)).subscribe( test=>
-      {
+    this.alertService.confirmThis("Êtes-vous sur de vouloir supprimer la tva ?", function () {
+      that.tvaService.deleteTva(tva).pipe(take(1)).subscribe(test => {
         var index = that.tvas.indexOf(tva);
         that.tvas.splice(index, 1);
-        that.messageService.add(message);        
+        that.messageService.add(message);
       }
-    );
-    that.selectedTva=null;
-    },function(){
+      );
+      that.selectedTva = null;
+    }, function () {
 
     });
-    
+
   }
 
-  updateDataForm(selectedTva:TvaI):void{
-    this.updateForm.patchValue({taux:selectedTva.taux});
+  updateDataForm(selectedTva: TvaI): void {
+    this.updateForm.patchValue({ taux: selectedTva.taux });
     this.updateModal.show();
-    this.selectedTva=selectedTva;
+    this.selectedTva = selectedTva;
   }
 
   onUpdate(): void {
-    const message:MessageI={content:'La modification a été enregistrée',level:'Info'};
-    this.selectedTva.taux=this.updateForm.get("taux").value;
+    const message: MessageI = { content: 'La modification a été enregistrée', level: 'Info' };
+    this.selectedTva.taux = this.updateForm.get("taux").value;
     this.tvaService.updateTva(this.selectedTva)
-      .pipe(take(1)).subscribe(item=>{
+      .pipe(take(1)).subscribe(item => {
         this.messageService.add(message);
         this.updateModal.hide();
       });
   }
 
   getTvas(): void {
-    this.tvaService.getTvas().pipe(take(1)).subscribe(tvas => this.tvas=tvas);
+    this.tvaService.getTvas().pipe(take(1)).subscribe(tvas => this.tvas = tvas);
   }
 
-  constructor(private tvaService:TvaService, private alertService: AlertService, private messageService:MessageService,private expireService:ExpireService) { }
+  constructor(private tvaService: TvaService, private alertService: AlertService, private messageService: MessageService, private expireService: ExpireService) { }
 
   ngOnInit(): void {
     this.getTvas();
